pyproject.toml
[tool.poetry]
name = "sc-dl"
version = "0.1.0"
description = "A simple CLI to download all of a SoundCloud user's liked tracks."
authors = ["Jason Antwi-Appah <hey@jasonaa.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
click = "^8.0.1"
python_dotenv = "^0.18.0"
soundcloud = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
